refactor: FormRequestを具象クラスではなくインタフェースに分離して、抽象的な扱いができるように修正することで、交差型として扱えるようになった
